% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{adj2str}
\alias{adj2str}
\title{Coerce adjacency matrix to a string}
\usage{
adj2str(adj.matrix)
}
\arguments{
\item{adj.matrix}{The pxp binary matrix representing the graph.}
}
\value{
The converted string.
}
\description{
This is a converter from the the adjacency matrix that represent a graph
to the default string of the sample.graphs vector, representing the upper
triangular matrix shrinked into a string of length (p-1)(p-2)/2.
}
