% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{create_structure}
\alias{create_structure}
\title{Data preprocessing for the FGM Hybrid Gibbs Sampler}
\usage{
create_structure(n = 351, p = 40, wave.indexes = c(1:235))
}
\arguments{
\item{n}{Number of curves which will be analysed. Default value is \code{351}.}

\item{p}{Dimension of the B-spline basis. Default value is \code{40}}

\item{wave.indexes}{Wavelengths range to analyse, expressed in terms of index.
Default value is the whole range of \code{235} wavelengths.}
}
\value{
A list with the re-arranged parameters provided in input.
}
\description{
This function pre-process the data stored in the \code{purees} data frame
(see \code{\link{purees}} for more information) and creates a data structure
which will be useful for the sampling algorithm (i.e. \code{\link{FGM_sampler}}).
}
