% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{FDR_analysis}
\alias{FDR_analysis}
\title{Bayesian FDR Analysis}
\usage{
FDR_analysis(PL, tol = seq(0.1, 1, by = 0.05), min_rate = 0.05)
}
\arguments{
\item{PL}{The posterior probability of inclusion of each link according to \code{\link{FGM_sampler}}.
It should be an upper diagonal matrix.}

\item{tol}{The sequence of thresholds, used to select a graph truncating PL at that value.}

\item{min_rate}{The upper bound for the false discoveries.}
}
\value{
a list of 3 elements:
\itemize{
  \item{\strong{best_threshold}, the best value of \code{tol} according to this analysis.}
  \item{\strong{best_truncated_graph}, the proposed posterior graph according to the analysis.}
  \item{\strong{best_distance}, the best FDR computed.}
}
}
\description{
Given the plinks matrix, this utility computes the False Discovery Rate Index,
forcing the false discovery rates to be less than \code{0.05}-
}
