\name{FGM-package}
\alias{FGM-package}
\docType{package}
\title{
Floating Grid Method
}
\description{
FGM is a spatially restricted permutation technique. The package contains both low level functions to produce permuted datasets as well as higher level functions to perform spatially restricted permutation tests and functions to summarize and plot the results.
}
\details{
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Reinder Radersma & Ben Sheldon
}
\references{
Reinder Radersma & Ben C. Sheldon, 2014. A new permutation test to explore and control for spatial autocorrelation in point pattern data. Manuscript submitted to MEE 
}
\keyword{ package }
\examples{

  }