% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{FGM}
\alias{FGM}
\title{FGM object constructor}
\usage{
FGM(X.axis, Y.axis, basis = "Spline", dim.basis)
}
\arguments{
\item{X.axis}{An r-dimensional vector that contains the abscissa points in which the data are evaluated.}

\item{Y.axis}{This can be both an r-dimensional vector f(X) or an n x p data frame which stores, in each
row, the values \ifelse{html}{\out{Y<sub>i</sub>=f(X<sub>i</sub>), i=1:n}}{\eqn{Y_i = f(X_i),~i=1:n}}.}

\item{basis}{A string that declares which type of basis needs to be used in order to embed data in their
functional framework. Right now, only "Spline" is available, ad uses a b-spline basis of third order.}

\item{dim.basis}{Integer, the dimension of the spline basis.}
}
\value{
A list of the elements, i.e.
\itemize{
  \item{\strong{data}, the dataframe provided in Y.axis.}
  \item{\strong{dims}, a named vector collecting the dimensions of the input data, i.e. n, number of curves, r, number of gridpoints and p, the basis dimension.}
  \item{\strong{init}, the \code{init} list that can be used as input in \code{\link{FGM_sampler}}.}
}
}
\description{
This utility recieves in input functional data defined in the format (X,Y), where X are the absciasse
at which the functions are evaluated and Y are the values of the functions in X. For a clarifying example,
see the format in which the \code{\link{purees}} dataset is stored in this package. It returns an FGM
object that collects useful information and helps the usage of these data in \code{\link{FGM_sampler}}.
}
