% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{FGM_sampler}
\alias{FGM_sampler}
\title{Sampling from the posterior distribution of the Functional Graphical Model}
\usage{
FGM_sampler(data, niter, nburn, thin, thinG, init, hyper = NULL)
}
\arguments{
\item{data}{The data that needs to be fitted through this model.}

\item{niter}{Number of iterations for the sampler.}

\item{nburn}{Number of burn-in iterations.}

\item{thin}{Thining parameter.}

\item{thinG}{Thining parameter for the joint precision-graph distribution.}

\item{init}{List that collects the starting point of the Gibbs Sampler Chains.
\itemize{
  \item{\strong{basemat}, the B-spline rxp basis matrix.}
  \item{\strong{G0}, starting graph (See \code{\link{bdgraph}} function for more information about the \code{g.start} input).}
  \item{\strong{Beta0}, list of p-dimensional vector, starting point for the \eqn{\beta} chains.}
  \item{\strong{mu0}, p-dimensional vector, starting point of the \eqn{\mu} chain.}
  \item{\strong{K0}, pxp Precision matrix, starting point for the \eqn{K} chain.}
  \item{\strong{tau_eps0}, real, starting point for the \ifelse{html}{\out{<em>&tau;<sub>&epsilon;</sub></em>}}{\eqn{\tau_\epsilon}} chain.}
}}

\item{hyper}{List that collects all the needed hyperparameters for the model. If nothing is passed, hyperparameters are set at their default values.
\itemize{
  \item{\strong{a_tau_eps}, first parameter of gamma prior for \ifelse{html}{\out{<em>&tau;<sub>&epsilon;</sub></em>}}{\eqn{\tau_\epsilon}}. \cr Default value: \eqn{2*10}.}
  \item{\strong{b_tau_eps}, second parameter of gamma prior for \ifelse{html}{\out{<em>&tau;<sub>&epsilon;</sub></em>}}{\eqn{\tau_\epsilon}}. \cr Default value: \eqn{2*0.001}.}
  \item{\strong{sigma_mu}, mean for the Gaussian prior for \eqn{\mu}. \cr Default value: \eqn{100}.}
  \item{\strong{d0}, degrees of freedom of the G-Wishart prior for the precision matrix. \cr Default value: \eqn{3}.}
  \item{\strong{gprior}, prior probability of each link of the graph (See \code{\link{bdgraph}} function for more information about the \code{g.prior} input). \cr Default value: \eqn{0.5}.}
}}
}
\value{
An \eqn{n+4}-item list with the following components:
\itemize{
  \item{\strong{beta_i} with \eqn{i=1:n}, the first n components are \code{(iter_to_store x p)} matrices storing the draws from the posterior distribution for all the \ifelse{html}{\out{<em>&beta;<sub>i</sub></em>}}{\eqn{\{\beta_{i}\}_{i}}}.}
  \item{\strong{mu}, an \code{(iter_to_store x p)} matrix collecting the \ifelse{html}{\out{<em>&mu;</em>}}{\eqn{\mu}} draws from the posterior distribution.}
  \item{\strong{tau_eps}, an \code{(iter_to_store)} vector collecting the \ifelse{html}{\out{<em>&tau;<sub>&epsilon;</sub></em>}}{\eqn{\tau_{\epsilon}}} draws from the posterior distribution.}
  \item{\strong{bdobject}, a \code{bdgraph} object (see \code{\link{bdgraph}} for details).}
  \item{\strong{K}, an \code{(iter_to_storeG)} vector collecting the \eqn{K} draws from the posterior distribution.}
}
}
\description{
This utility implements an hybrid Gibbs Sampler strategy in order to sample
the posterior distribution of the parameters \ifelse{html}{\out{<em>&beta;<sub>i</sub>, &mu;, K, G, &tau;<sub>&epsilon;</sub></em>}}{\eqn{\{\beta_{i}\}_{i}, \mu, K, G, \tau_{\epsilon}}}
of the Functional Graphical Model for the spectrometric data analysis performed.
}
