% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_things.R
\name{FGM_sim}
\alias{FGM_sim}
\title{Functional Graphical Model simulation}
\usage{
FGM_sim(S0, X.axis, dims, basis = "Spline", method, sparsity = 0.3, type = 4)
}
\arguments{
\item{S0}{Natural number, the seed for the random generator for the data.}

\item{X.axis}{An r-dimensional vector that contains the abscissa points in which the data are evaluated.}

\item{dims}{A named vector collecting the dimensions of the input data, i.e. n, number of curves, r, number of gridpoints and p, the basis dimension.}

\item{basis}{A string that declares which type of basis needs to be used in order to embed data in their
functional framework. Right now, only "Spline" is available, ad uses a b-spline basis of third order.}

\item{sparsity}{\eqn{[0, 1]}-ranged parameter that indicates how sparse the precision matrix of the
functional graphical model is required. Default value is \eqn{0.3}.}

\item{type}{Select the desider Peterson graph/precision matrix couple. See \code{\link{KG_sim}}.}
}
\value{
A list containing all the simulated data for the Functional Graphical Model in the form of an FGM object and a list with the parameters of simulation, that can be useful for debugging purposes.
}
\description{
This utility simulates data from the Functional Graphical Model under analysis. Since the seed of
the random generator can be fixed, this allows multiple replications of the same simulated data
in order to perform multiple check on the performed analysis.
}
