% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{hamming_distance_analysis}
\alias{hamming_distance_analysis}
\title{Bayesian Hamming Distance Analysis}
\usage{
hamming_distance_analysis(PL, G_true, tol = seq(0.1, 1, by = 0.05))
}
\arguments{
\item{PL}{The posterior probability of inclusion of each link according to \code{\link{FGM_sampler}}.
It should be an upper diagonal matrix.}

\item{G_true}{The true graph.}

\item{tol}{The sequence of thresholds, used to select a graph truncating PL at that value.}
}
\value{
A list of 3 elements:
\itemize{
  \item{\strong{best_threshold}, the best value of \code{tol} according to this analysis.}
  \item{\strong{best_truncated_graph}, the proposed posterior graph according to the analysis.}
  \item{\strong{best_distance}, the best value of the hamming distance, the one corresponding to \code{best_threshold}.}
}
}
\description{
Given the true graph and a plinks matrix, this utiliry computes the hamming distances between
the real graph and the truncated ones.
}
