% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_things.R
\name{KG_sim}
\alias{KG_sim}
\title{Graph and Precision matrix joint simulation}
\usage{
KG_sim(method, p, sparsity, S0, type)
}
\arguments{
\item{method}{The String identifying which method will be used to simulate the graph.}

\item{p}{The B-spline basis dimension. Is required to be multiple of 4, if \code{method="Peterson"}.}

\item{sparsity}{The [0, 1]-ranged parameter that indicates how sparse the precision matrix of the
functional graphical model is required. Used in case \code{method='Bernoulli'}, otherwise discarded.}

\item{S0}{Natural number, the seed for the random generator for the data.}

\item{type}{A natural number between 1 and 4, which selects which Peterson matrix will be used
for the sampling. See \code{\link{Peterson_p16}} and \code{\link{Peterson_p40}}.}
}
\value{
A list containing the sampled graph and the induced precision matrix.
}
\description{
This function provides a graph simulation at a fixed seed for a random generator, according
to different methods. In particular, three methods are here implemented:
\itemize{
  \item{\code{Bernoulli} associates at each link generation a Bernoulli trial.}
  \item{\code{Cluster} induces a graph with a blocking structure, according to the sparsity parameter.}
  \item{\code{Peterson} this method exploits literature precisions matrices commonly used for simulations.
  For further references, read, for instance, \url{https://www.researchgate.net/publication/261020849}.}
}
}
