% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_manipulation.R
\name{mean_beta_per_component}
\alias{mean_beta_per_component}
\title{Compute the component-wise mean for beta}
\usage{
mean_beta_per_component(beta)
}
\arguments{
\item{beta}{The list of beta coming from select_all_beta function.}
}
\value{
A p-dimensional vector containing the mean value of the
j-th component of beta.
}
\description{
This function compute the mean value of beta per component.
}
