% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{misclass_analysis}
\alias{misclass_analysis}
\title{Bayesian Sensitivity Analysis}
\usage{
misclass_analysis(PL, G_true, tol = seq(0.1, 1, by = 0.05), ROC = FALSE)
}
\arguments{
\item{PL}{The posterior probability of inclusion of each link according to \code{\link{FGM_sampler}}.
It should be an upper diagonal matrix.}

\item{G_true}{The true graph.}

\item{tol}{The sequence of thresholds, used to select a graph truncating PL at that value.}

\item{ROC}{Boolean parameter. If \code{TRUE}, the plot the ROC curve is showed.}
}
\value{
A list of 5 elements:
\itemize{
  \item{\strong{all_confusion}, a list with all the confusion matrices, one for each tol value.}
  \item{\strong{best_threshold}, the best value of \code{tol} according to this analysis.}
  \item{\strong{best_confusion}, the best confusion matrix, the one corresponding to \code{best_threshold}.}
  \item{\strong{best_truncated_graph}, the proposed posterior graph according to the analysis.}
  \item{\strong{AUC}, the AUC Index corresponding to the best value selected.}
}
}
\description{
Given the true graph and a plinks matrix, this utility computes the confusion matrices, plot the ROC curve (if required)
and the AUC index, selects the best threshold according to the number of misclassified links
and also return the best graph according to the abovementioned analysis.
}
