% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Peterson_p40}
\alias{Peterson_p40}
\title{Peterson's Graphs and Precision Matrices}
\format{
A list of four elements, each of the is a list containing:
\describe{
  \item{G}{The Peterson's graph suggested in the reference paper.}
  \item{K}{A pxp precision matrix iduced by G.}
}
}
\source{
\url{https://www.researchgate.net/publication/261020849}
}
\usage{
data("Peterson_p16") or data("Peterson_p40")
}
\description{
This dataset contains a list of four elements, which contains the graph and precision matrices
of dimension pxp suggested by Peterson et al. (2015) in order to perform simulations on graphical models.
}
\keyword{datasets}
