% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purees}
\alias{purees}
\title{Absorbance mid-IR spectra of 983 fruit purées.}
\format{
A list that contains two elements. \strong{wavelengths} contains the above-mentioned
vector of wavelengths, while \strong{data} is a data frame with 983 rows and 236 variables:
\describe{
  \item{Group}{Factor variable identifying if the specturm comes from a strawberry purees or not.}
  \item{W_899.327}{Absorbance value registered at 899.327 nm. All other variables have similar meaning.}
}
}
\source{
\url{https://csr.quadram.ac.uk/example-datasets-for-download/}
}
\usage{
data("purees")
}
\description{
This dataset contains the absorbance spectra of two different classes of fruit purées, evaluated
in 235 different wavelenghts. Among them, we are able to distinguish data coming from pure strawberry
purees and from adulterated purées, obtained through a mixture of different fruits. This dataset is
in the proper format for the creation of an FGM object. See \code{\link{FGM}} for details.
}
\keyword{datasets}
