% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{BDgraph sampling from multivariate normal}
\usage{
rmvnorm(
  n = 10,
  mean = rep(0, length = ncol(sigma)),
  sigma = diag(length(mean))
)
}
\description{
BDgraph sampling from multivariate normal
}
