% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_JAGS.R
\name{select_all_beta}
\alias{select_all_beta}
\title{Extract the whole beta posterior sampling.}
\usage{
select_all_beta(output, n.curves)
}
\arguments{
\item{output}{The JAGS output.}

\item{n.curves}{The number of curves for which we've executed the sampling
strategy with JAGS.}
}
\value{
A list in which the [[i]]-th element is the dataframe.
referring to the i-th curve.
}
\description{
Function built specifically to group all the beta posterior samplings coming
from JAGS by sample unit. It creates a list in which the [[i]]-th element is
a dataframe that collects the beta samplings related to the i-th curves.
}
