% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_JAGS.R
\name{select_posterior_FM}
\alias{select_posterior_FM}
\title{Extract selected JAGS output}
\usage{
select_posterior_FM(output, param, index = NULL)
}
\arguments{
\item{output}{The JAGS output.}

\item{index}{Required in param="beta", select all the betas related to the i-th
sample. Default vaule is \code{NULL}.}
}
\value{
A dataframe containing the required information as columns.
}
\description{
Useful function that extracts the required information in the form of a dataframe.
Use this to extract a specific beta sample, mu, tau ad tau_eps JAGS samplings from
the posterior.
}
