% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smooth_FGM}
\alias{smooth_FGM}
\title{Bayesian smoothing for the Functional (Graphical) Model}
\usage{
smooth_FGM(beta, basemat, alpha = 0.1, band = FALSE)
}
\arguments{
\item{beta}{A list of length \eqn{n}, with all the sampled values for each beta coefficient. One can either
pass the post-processed output for the Functional Model coming from \code{\link{select_all_beta}} or
the first \eqn{n} elements of the output list for the Functional Graphical Model, coming from \code{\link{FGM_sampler}}.}

\item{basemat}{The design matrix given by the evaluation of the basis over the grid of points.}

\item{alpha}{A [0-1]-ranged parameter to be set in order to get the \ifelse{html}{\out{(1-&alpha;)\%}}{\eqn{(1-\alpha)\%}} credibility bands.}

\item{band}{Boolean, \code{FALSE} by default. If \code{TRUE}, the function also computes the \ifelse{html}{\out{(1-&alpha;)\%}}{\eqn{(1-\alpha)\%}}
credibility bands.}
}
\value{
In case band is \code{FALSE}, it returns a matrix with all the values of the smoothed curves.
If band is \code{TRUE}, it returns a list of three matrices collecting the lower, the mean and the upper bands values of the smoothing.
}
\description{
This function computes the smoothed curves and their credibility bands, if required.
}
