% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_manipulation.R
\name{SS_beta}
\alias{SS_beta}
\title{Compute the sum of squares of beta}
\usage{
SS_beta(beta, mu = NULL)
}
\arguments{
\item{beta}{The list of beta coming from select_all_beta function.}

\item{mu}{The p-dimensional vector of the mean.}
}
\value{
A p-dimensional vector a for which a[j] = \ifelse{html}{\out{&Sigma;<sub>i</sub>(&beta;<sub>ij</sub> - &mu;<sub>j</sub>)<sup>2</sup>}}{\eqn{\sum_{i}\left(\beta_{ij}-\mu_{j}\right)^2}}.
}
\description{
This utility simply compute the sum of sqares of the betas related to the
i-th curve. Useful tool in out custom GS.
}
