% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{str2adj}
\alias{str2adj}
\title{Conversion to adjacency matrix}
\usage{
str2adj(p, graph.str)
}
\arguments{
\item{p}{Size of the graph, i.e. the number of nodes.}

\item{graph.str}{Binary string to convert}
}
\value{
A pxp binary matrix representing the graph in input.
}
\description{
This is a converter from the default string of the sample.graphs vector
to its proper adjacency matrix, in the form of a binary matrix.
}
