% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{str2K}
\alias{str2K}
\title{Conversion to precision matrix}
\usage{
str2K(p, K.str)
}
\arguments{
\item{p}{Size of the graph, i.e. the number of nodes.}

\item{K.str}{The comma-separated string containing the values of the upper
triangular matrix (diagonal included)}
}
\value{
The converted pxp spd precision matrix.
}
\description{
This is a converter from a comma-separated string representing the upper
triangular precision matrix (diagonal included) to the equivalent pxp positive
definite precision matrix. Useful since precision matrices are stored as strings
in order to reduce space.
}
