% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrawberryPurees}
\alias{StrawberryPurees}
\title{Absorbance mid-IR spectra of 351 strawberry purées.}
\format{
An object of class \code{data.frame} with 351 rows and 235 columns.
}
\usage{
data("StrawberryPurees")
}
\description{
This is a shortcut to load only a part of the dataset described in \code{\link{purees}}.
This dataset contains only the 351 absorbance spectra of strawberry purées, measured at 235 different wavelenghts.
Differently from \code{\link{purees}}, this file contains only the data.frame of the measurments. The corresponding
wavelengths can be loaded using \code{\link{StrawberryWavelengths}}.
}
\keyword{datasets}
