% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrawberryWavelengths}
\alias{StrawberryWavelengths}
\title{Grid of wavelengths where absorbance spectra are measured.}
\format{
An object of class \code{data.frame} with 235 rows and 1 columns.
}
\usage{
data("StrawberryWavelengths")
}
\description{
This is a shortcut to load only a part of the dataset described in \code{\link{purees}}.
This dataset contains a vector of 235 elements representing the grid where the absorbance spectra availabe in \code{\link{StrawberryPurees}}
are measured.
}
\keyword{datasets}
