% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_FM}
\alias{summary_FM}
\title{Summarizing posterior distribution of the FM model}
\usage{
summary_FM(output, n.curves, alpha = 0.05)
}
\arguments{
\item{output}{The JAGS output.}

\item{n.curves}{The number of curves for which we've computed the
posterior distribution.}

\item{alpha}{Real number used to compute the (1-\eqn{\alpha}) credible intervals.
Default value is 0.05, hence we compute the 95\% CI.}
}
\value{
A total of five plots which represent, in order, the mean
value of each beta in the form of a matrix, the correlation
plot of the beta coefficients, the (1-\eqn{\alpha})\% credible intervals
for the mu vector, the tau vector and the tau_eps parameter.
}
\description{
This utility provides useful plots for an overall summary of the posterior
distribution of all parameters which come from the JAGS output of the Functional Model
}
